pyproject.toml
[tool.poetry]
name = "zasek"
version = "0.1.0"
description = ""
authors = ["otters-team"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.1"
djangorestframework = "^3.9"
psycopg2 = "^2.7"
django-rest-auth = "^0.9.3"
django-allauth = "^0.38.0"
drf-yasg = "^1.13"
psycopg2-binary = "^2.7"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
